// src/app/shared/model/booking/booking-request.ts
export class BookingRequest {
  id: number;
  fullName: string;
  address: string;
  email: string;
  phone: string;
  description: string;
  id_user: number;
  id_house: number;
  price: number;
  costsIncurred: number;
  discount: number;
  deposit: boolean;
  dateIn: string;
  dateOut: string;
  identityCard: string;
  id_creator: number;
  }